refactor: prepare day12 part2
As we will need to find the better shortest path starting from a
'a', the easiest way will be to inverse the search and start from
the end and stop as soon as we found a 'a'.

So prepare the code for that, for the solution 1 instead of
searching a path from start to end, we will search a path from end
to start (we just need to inverse the call to `isWalkable`).