feat(solution): day16 part2
Generate all paths and see if are two paths not having the same
valves, keep the best combination.

We could also have rewritten the DFS to have two runners, but it
seems to be quicker in term of implementation like that, and I
already spent too much time on this problem...

Maybe I will revisit another time.